Adding test scripts and updating gitshelve
Adding virtual environment setup scripts.
Adding runTests, which will make python 2 and 3 environments if the executables are in the path.  Once the environments are made, it will run the tests in both environments and generate coverage reports.
Fixing t_gitshelve to include the original unit tests.
Adding put and get functions so the original unit tests work.
Reducing the complexity of the git function so the sonar warnings are reduced.